Launcher3: Fix upstream bug with dots enabled (#5167)
https://github.com/GrapheneOS/os-issue-tracker/issues/2601

I did not find any issue reported upstream, but I can confirm that this
bug exists on Moto App Launcher (the launcher on Motorola's devices),
which seems to imply this is a bug in AOSP itself. I had asked
agnostic-apollo about this is a bug, and he said that this is a bug in
Android's libraries itself

neobuddy89: Update error message.

Change-Id: Icd41b0a8b3c71706f37ef01a81caf81c502d3ff2
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>